blog/pyproject.toml
[project]
name = "django-blog"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-storages[s3]>=1.14.6",
    "Markdown>=3.10.2",
    "celery[redis]>=5.4",
    "nh3>=0.3.3",
    "pillow>=12.1.0",
    "pillow-heif>=1.3.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.2.1",
    "rawpy>=0.26.1",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]
